+++ src/version.py
@@ -1,1 +1,1 @@
-__version__ = "bc7bc40e69c5b69e7cb376cd823528ded5487035"
+__version__ = "306ef9e97c54c9cb7ec3d87f640e00a1a5690df7"
