+++ src/vision/sdxl_turbo.py
